+++ ee/modules/650-runtime-audit-engine/images/rules-loader/hooks/lib/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2023 Flant JSC
+# Licensed under the Deckhouse Platform Enterprise Edition (EE) license. See https://github.com/deckhouse/deckhouse/blob/main/ee/LICENSE
